+++ public/index.html
@@ -9,9 +9,6 @@
       name="description"
       content="Web site created using create-react-app"
     />
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
-
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
     <title>21세기전파상 파트너어드민</title>
   </head>
   <body>
