+++ src/components/common/Paging.jsx
@@ -1,0 +1,24 @@
+import React from "react";
+import Pagination from "@mui/material/Pagination";
+
+const Paging = ({ totalcnt, onChangepage, limit, currentpage }) => {
+  let pageCnt = totalcnt % limit;
+  if (pageCnt === 0) {
+    pageCnt = parseInt(totalcnt / limit);
+  } else {
+    pageCnt = parseInt(totalcnt / limit) + 1;
+  }
+
+  return (
+    <div>
+      <Pagination
+        count={pageCnt ? pageCnt : 1}
+        page={parseInt(currentpage)}
+        onChange={onChangepage}
+        showFirstButton
+        showLastButton
+      />
+    </div>
+  );
+};
+export default Paging;
